Moved stuff to player and team stats.
