+++ carball/analysis/hit_detection/base_hit.py
@@ -43,7 +43,6 @@ class BaseHit:
         hit_creation_time = time.time()
         logger.info('time to get get frame_numbers: %s', (hit_creation_time - start_time) * 1000)
 
-
         hit_frames = data_frame.loc[hit_frame_numbers, (slice(None), positional_columns)]
         player_displacements = {player.name: get_player_ball_displacements(hit_frames, player.name)
                                 for player in game.players}
