+++ carball/analysis/stats/stats_list.py
@@ -1,18 +1,19 @@
 from typing import List
 
-from ...analysis.stats.tendencies.relative_position_tendencies import RelativeTendencies
-from ...analysis.stats.tendencies.team_tendencies import TeamTendencies
-from ...analysis.stats.tendencies.speed_tendencies import SpeedTendencies
-from ...analysis.stats.tendencies.averages import Averages
-from ...analysis.stats.possession.ball_distances import BallDistanceStat
-from ...analysis.stats.tendencies.hit_counts import HitCountStat
-from ...analysis.stats.ball_forward.distance_hit_ball_forward import DistanceStats
-from ...analysis.stats.boost.boost import BoostStat
-from ...analysis.stats.possession.possession import PossessionStat
-from ...analysis.stats.possession.turnovers import TurnoverStat
-from ...analysis.stats.stats import BaseStat, HitStat
-from ...analysis.stats.tendencies.positional_tendencies import PositionalTendencies
-from ...analysis.stats.controls.controls import ControlsStat
+from carball.analysis.stats.possession.per_possession import PerPossessionStat
+from carball.analysis.stats.ball_forward.distance_hit_ball_forward import DistanceStats
+from carball.analysis.stats.boost.boost import BoostStat
+from carball.analysis.stats.controls.controls import ControlsStat
+from carball.analysis.stats.possession.ball_distances import BallDistanceStat
+from carball.analysis.stats.possession.possession import PossessionStat
+from carball.analysis.stats.possession.turnovers import TurnoverStat
+from carball.analysis.stats.stats import BaseStat, HitStat
+from carball.analysis.stats.tendencies.averages import Averages
+from carball.analysis.stats.tendencies.hit_counts import HitCountStat
+from carball.analysis.stats.tendencies.positional_tendencies import PositionalTendencies
+from carball.analysis.stats.tendencies.relative_position_tendencies import RelativeTendencies
+from carball.analysis.stats.tendencies.speed_tendencies import SpeedTendencies
+from carball.analysis.stats.tendencies.team_tendencies import TeamTendencies
 
 
 class StatsList:
@@ -28,7 +29,8 @@ class StatsList:
                 Averages(),
                 BallDistanceStat(),
                 ControlsStat(),
-                SpeedTendencies()
+                SpeedTendencies(),
+                PerPossessionStat(),
                 ]
 
     @staticmethod
@@ -36,7 +38,9 @@ class StatsList:
         """These are stats that end up being assigned to a specific team"""
         return [PossessionStat(),
                 TeamTendencies(),
-                RelativeTendencies()]
+                RelativeTendencies(),
+                PerPossessionStat(),
+                ]
 
     @staticmethod
     def get_general_stats() ->List[BaseStat]:
